Crate m5dial_bsp

M5 Dial Board Support Package (BSP)

This crates is a Board support package for the M5 Dial.

API Documentation

Feature list/roadmap:

Screen driver (GC9A01)

Rotary encoder

Button

Power management (Shutdown)

Touch-screen (FT3267)

Buzzer

Real-time clock (BM8563)

Port A (I2C)

Port B

Upgrade to be compatible to esp-hal v1.0.0

How to use

First generate a base project using esp-generate as described in its documentation. Then, add this crate to your Cargo.toml:

[dependencies]
....
m5dial-bsp= "0.5.0"
....

In your main function then initialize this hall and use it:

let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
let peripherals = esp_hal::init(config);

let mut display = m5dial_bsp::get_screen!(peripherals);
let mut encoder = m5dial_bsp::get_encoder!(peripherals);

let mut board = m5dial_bsp::board_init!(peripherals);

For more information, please refer to the API Documentation. (Don’t refers to the doc on docs.rs, as they fail to compile due to the lack of ESP specific toolchain.)

Examples

Examples are located in the examples folder. Here is an index:

screen_counter.rs: Demonstrate the rotary encoder usage.

screen_counter_irq.rs: Demonstrate the rotary encoder usage, using interrupts.

touch.rs: Demonstrate the touchscreen.

License

MIT license (LICENSE or http://opensource.org/licenses/MIT)

Contributions

… are welcome !

Credit

This crates rely on many crates of the embedded Rust working group, ESP support crates from Expressif, embedded-graphics and device drivers such as :

rotary-encoder-hal

gc9c01 display driver

Re-exports

pub use bsp::M5DialBsp;

Modules

bsp

M5Dial Board Support Package

buzzer

ft3267

RT3267 Touch driver

Macros

board_init

Initialize board peripherals from ESP32 peripherals.

get_buzzer

Get the buzzer

get_encoder

Get rottary encoder

get_port_a_i2c

Get the Port.A external I2C bus

get_port_b_in

Get PORT.B input (white wire)

get_port_b_out

Get PORT.B output (yellow wire)

get_screen

Get screen controller

get_touch

Get the touch screen controller