m5dial_bsp

Module bsp

M5Dial Board Support Package

Re-exports

pub use crate::ft3267::Ft3267;

pub use crate::ft3267::TouchPoint;

pub use crate::buzzer::Buzzer;

Macros

error

Structs

Blocking

Marker type signalling that a driver is initialized in blocking mode.

Brightness

Struct that holds display brightness

BufferedGraphics

Buffered Graphic Implementation

DefaultPhase

Default implementation of Phase.

Delay

Delay driver, using [Instant].

DisplayResolution240x240

Screen Definition Resolution 240 x 240

EspI2C

I2C driver

ExclusiveDevice

[SpiDevice] implementation with exclusive access to the bus (not shared).

Gc9a01

Gc9a01 Driver

I2cConfig

I2C driver configuration

Input

Digital input.

InputConfig

Input pin configuration.

M5DialBsp

Holds the board peripherals

NoDelay

Dummy DelayNs implementation that panics on use.

Output

Push-pull digital output.

OutputConfig

Output pin configuration.

Rate

Represents a rate or frequency of events.

Rmt

RMT Instance

Rotary

Holds current/old state and both InputPin

SPIDisplayInterface

SPI Interfaces for the screen

SPIInterface

SPI display interface.

Spi

SPI peripheral driver

SpiConfig

SPI peripheral configuration

TouchControl

Enums

DisplayRotation

Screen Rotation

Level

Digital input or output level.

Mode

SPI communication modes, defined by clock polarity (CPOL) and clock phase (CPHA).

Pull

Pull setting for a GPIO.

Traits

DisplayConfiguration

DisplayDefinition

Screen information

OutputPin

Trait implemented by pins which can be used as outputs.

WriteOnlyDataCommand

This trait implements a write-only interface for a display which has separate data and command modes. It is the responsibility of implementations to activate the correct mode in their implementation when corresponding method is called.

Type Aliases

M5DialDisplay

Define a type alias for the display

M5DialEncoder