m5dial_bsp::bsp

Struct EspI2C

pub struct EspI2C<'d, Dm>
where
    Dm: DriverMode,
{ /* private fields */ }

I2C driver

Example

use esp_hal::i2c::master::{Config, I2c};
let mut i2c = I2c::new(peripherals.I2C0, Config::default())?
    .with_sda(peripherals.GPIO1)
    .with_scl(peripherals.GPIO2);

let mut data = [0u8; 22];
i2c.write_read(DEVICE_ADDR, &[0xaa], &mut data)?;

Implementations

impl<'d> I2c<'d, Blocking>

pub fn new( i2c: impl Instance + 'd, config: Config, ) -> Result<I2c<'d, Blocking>, ConfigError>

Create a new I2C instance.

Errors

A [ConfigError] variant will be returned if bus frequency or timeout passed in config is invalid.

Example

use esp_hal::i2c::master::{Config, I2c};
let i2c = I2c::new(peripherals.I2C0, Config::default())?
    .with_sda(peripherals.GPIO1)
    .with_scl(peripherals.GPIO2);

pub fn into_async(self) -> I2c<'d, Async>

Reconfigures the driver to operate in [Async] mode.

See the [Async] documentation for an example on how to use this method.

pub fn set_interrupt_handler(&mut self, handler: InterruptHandler)

Registers an interrupt handler for the peripheral on the current core.

Note that this will replace any previously registered interrupt handlers.

You can restore the default/unhandled interrupt handler by passing [DEFAULT_INTERRUPT_HANDLER][crate::interrupt::DEFAULT_INTERRUPT_HANDLER].

Panics

Panics if passed interrupt handler is invalid (e.g. has priority None)

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn listen(&mut self, interrupts: impl Into<EnumSet<Event>>)

Listen for the given interrupts

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn unlisten(&mut self, interrupts: impl Into<EnumSet<Event>>)

Unlisten the given interrupts

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn interrupts(&mut self) -> EnumSet<Event>

Gets asserted interrupts

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn clear_interrupts(&mut self, interrupts: EnumSet<Event>)

Resets asserted interrupts

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

impl<'d> I2c<'d, Async>

pub fn into_blocking(self) -> I2c<'d, Blocking>

Reconfigures the driver to operate in Blocking mode.

See the Blocking documentation for an example on how to use this method.

pub async fn write_async<A>( &mut self, address: A, buffer: &[u8], ) -> Result<(), Error>
where A: Into<I2cAddress>,

Writes bytes to slave with given address.

Note that dropping the returned Future will abort the transfer, but doing so will block while the driver is finishing clearing and releasing the bus.

Example

use esp_hal::i2c::master::{Config, I2c};
const DEVICE_ADDR: u8 = 0x77;
let mut i2c = I2c::new(peripherals.I2C0, Config::default())?
    .with_sda(peripherals.GPIO1)
    .with_scl(peripherals.GPIO2)
    .into_async();

i2c.write_async(DEVICE_ADDR, &[0xaa]).await?;

pub async fn read_async<A>( &mut self, address: A, buffer: &mut [u8], ) -> Result<(), Error>
where A: Into<I2cAddress>,

Reads enough bytes from slave with address to fill buffer.

Note that dropping the returned Future will abort the transfer, but doing so will block while the driver is finishing clearing and releasing the bus.

Errors

The corresponding error variant from [Error] will be returned if the passed buffer has zero length.

Example

use esp_hal::i2c::master::{Config, I2c};
const DEVICE_ADDR: u8 = 0x77;
let mut i2c = I2c::new(peripherals.I2C0, Config::default())?
    .with_sda(peripherals.GPIO1)
    .with_scl(peripherals.GPIO2)
    .into_async();

let mut data = [0u8; 22];
i2c.read_async(DEVICE_ADDR, &mut data).await?;

pub async fn write_read_async<A>( &mut self, address: A, write_buffer: &[u8], read_buffer: &mut [u8], ) -> Result<(), Error>
where A: Into<I2cAddress>,

Writes bytes to slave with given address and then reads enough bytes to fill buffer in a single transaction.

Note that dropping the returned Future will abort the transfer, but doing so will block while the driver is finishing clearing and releasing the bus.

Errors

The corresponding error variant from [Error] will be returned if the passed buffer has zero length.

Example

use esp_hal::i2c::master::{Config, I2c};
const DEVICE_ADDR: u8 = 0x77;
let mut i2c = I2c::new(peripherals.I2C0, Config::default())?
    .with_sda(peripherals.GPIO1)
    .with_scl(peripherals.GPIO2)
    .into_async();

let mut data = [0u8; 22];
i2c.write_read_async(DEVICE_ADDR, &[0xaa], &mut data)
    .await?;

pub async fn transaction_async<'a, A>( &mut self, address: A, operations: impl IntoIterator<Item = &'a mut Operation<'a>>, ) -> Result<(), Error>
where A: Into<I2cAddress>,

Execute the provided operations on the I2C bus as a single transaction.

Note that dropping the returned Future will abort the transfer, but doing so will block while the driver is finishing clearing and releasing the bus.

Transaction contract:

Before executing the first operation an ST is sent automatically. This is followed by SAD+R/W as appropriate.

Data from adjacent operations of the same type are sent after each other without an SP or SR.

Between adjacent operations of a different type an SR and SAD+R/W is sent.

After executing the last operation an SP is sent automatically.

If the last operation is a Read the master does not send an acknowledge for the last byte.

ST = start condition

SAD+R/W = slave address followed by bit 1 to indicate reading or 0 to indicate writing

SR = repeated start condition

SP = stop condition

Errors

The corresponding error variant from [Error] will be returned if the buffer passed to an [Operation] has zero length.

Example

use esp_hal::i2c::master::{Config, I2c, Operation};
const DEVICE_ADDR: u8 = 0x77;
let mut i2c = I2c::new(peripherals.I2C0, Config::default())?
    .with_sda(peripherals.GPIO1)
    .with_scl(peripherals.GPIO2)
    .into_async();

let mut data = [0u8; 22];
i2c.transaction_async(
    DEVICE_ADDR,
    &mut [Operation::Write(&[0xaa]), Operation::Read(&mut data)],
)
.await?;

impl<'d, Dm> I2c<'d, Dm>
where Dm: DriverMode,

pub fn with_sda(self, sda: impl PeripheralOutput<'d>) -> I2c<'d, Dm>

Connect a pin to the I2C SDA signal.

This will replace previous pin assignments for this signal.

pub fn with_scl(self, scl: impl PeripheralOutput<'d>) -> I2c<'d, Dm>

Connect a pin to the I2C SCL signal.

This will replace previous pin assignments for this signal.

Example

use esp_hal::i2c::master::{Config, I2c};
const DEVICE_ADDR: u8 = 0x77;
let i2c = I2c::new(peripherals.I2C0, Config::default())?.with_scl(peripherals.GPIO2);

pub fn write<A>(&mut self, address: A, buffer: &[u8]) -> Result<(), Error>
where A: Into<I2cAddress>,

Writes bytes to slave with given address

Example

use esp_hal::i2c::master::{Config, I2c};
i2c.write(DEVICE_ADDR, &[0xaa])?;

pub fn read<A>(&mut self, address: A, buffer: &mut [u8]) -> Result<(), Error>
where A: Into<I2cAddress>,

Reads enough bytes from slave with address to fill buffer

Example

use esp_hal::i2c::master::{Config, I2c};
let mut data = [0u8; 22];
i2c.read(DEVICE_ADDR, &mut data)?;

Errors

The corresponding error variant from [Error] will be returned if the passed buffer has zero length.

pub fn write_read<A>( &mut self, address: A, write_buffer: &[u8], read_buffer: &mut [u8], ) -> Result<(), Error>
where A: Into<I2cAddress>,

Writes bytes to slave with given address and then reads enough bytes to fill buffer in a single transaction

Errors

The corresponding error variant from [Error] will be returned if the passed buffer has zero length.

Example

use esp_hal::i2c::master::{Config, I2c};
let mut data = [0u8; 22];
i2c.write_read(DEVICE_ADDR, &[0xaa], &mut data)?;

pub fn transaction<'a, A>( &mut self, address: A, operations: impl IntoIterator<Item = &'a mut Operation<'a>>, ) -> Result<(), Error>
where A: Into<I2cAddress>,

Execute the provided operations on the I2C bus.

Transaction contract:

Before executing the first operation an ST is sent automatically. This is followed by SAD+R/W as appropriate.

Data from adjacent operations of the same type are sent after each other without an SP or SR.

Between adjacent operations of a different type an SR and SAD+R/W is sent.

After executing the last operation an SP is sent automatically.

If the last operation is a Read the master does not send an acknowledge for the last byte.

ST = start condition

SAD+R/W = slave address followed by bit 1 to indicate reading or 0 to indicate writing

SR = repeated start condition

SP = stop condition

Example

use esp_hal::i2c::master::{Config, I2c, Operation};
let mut data = [0u8; 22];
i2c.transaction(
    DEVICE_ADDR,
    &mut [Operation::Write(&[0xaa]), Operation::Read(&mut data)],
)?;

Errors

The corresponding error variant from [Error] will be returned if the buffer passed to an [Operation] has zero length.

pub fn apply_config(&mut self, config: &Config) -> Result<(), ConfigError>

Applies a new configuration.

Errors

A [ConfigError] variant will be returned if bus frequency or timeout passed in config is invalid.

Example

use esp_hal::i2c::master::{Config, I2c};
let mut i2c = I2c::new(peripherals.I2C0, Config::default())?;

i2c.apply_config(&Config::default().with_frequency(Rate::from_khz(400)))?;