m5dial_bsp::bsp

Struct Gc9a01

pub struct Gc9a01<I, D, M>
where
    I: WriteOnlyDataCommand,
    D: DisplayDefinition,
{ /* private fields */ }

Gc9a01 Driver

Implementations

impl<I, D> Gc9a01<I, D, BasicMode>
where I: WriteOnlyDataCommand, D: DisplayDefinition,

pub fn clear(&mut self) -> Result<(), DisplayError>

Clear the display

Errors

This method may return an error if there are communication issues with the display.

pub fn set_pixels( &mut self, start: (u16, u16), end: (u16, u16), colors: &mut dyn Iterator<Item = u16>, ) -> Result<(), DisplayError>

Set the pixels directly to the hardware by setting the window from start to end based on the Iterator<Item = u16> provided.

This function does not protect the user input.

Errors

This method may return an error if there are communication issues with the display.

pub fn set_pixel( &mut self, x: u16, y: u16, value: u16, ) -> Result<(), DisplayError>

Set a pixel color at x and y coordinates directly through the hardware.

This function does not protect the user input.

Errors

This method may return an error if there are communication issues with the display.

impl<I, D> Gc9a01<I, D, BufferedGraphics<D>>
where I: WriteOnlyDataCommand, D: DisplayDefinition,

pub fn clear(&mut self)

Clear the display buffer NOTE: Must use flush to apply changes

pub fn fill(&mut self, color: u16)

pub fn flush(&mut self) -> Result<(), DisplayError>

Write the display buffer

Errors

This method may return an error if there are communication issues with the display.

pub fn set_pixels<T>( &mut self, start: (u16, u16), end: (u16, u16), colors: T, ) -> Result<(), DisplayError>
where T: IntoIterator<Item = u16>,

Set the pixels

Errors

This method may return an error if there are communication issues with the display. This method may return an error if there are an out of bounds error.

pub fn set_pixel(&mut self, x: u32, y: u32, value: u16)

Set a pixel color. If the X and Y coordinates are out of the bounds of the display, this method call is a noop

impl<I, D, M> Gc9a01<I, D, M>
where I: WriteOnlyDataCommand, D: DisplayDefinition,

pub fn reset<RST, DELAY>( &mut self, rst: &mut RST, delay: &mut DELAY, ) -> Result<(), <RST as ErrorType>::Error>
where RST: OutputPin, DELAY: DelayNs,

Reset the display.

Errors

See OutputPin definition for more information.

impl<I, D> Gc9a01<I, D, BasicMode>
where I: WriteOnlyDataCommand, D: DisplayDefinition,

pub fn new( interface: I, screen: D, screen_rotation: DisplayRotation, ) -> Gc9a01<I, D, BasicMode>

Create a basic Gc9a01 interface.

Use the into_buffed_graphics methods to enable more functionality.

impl<I, D, M> Gc9a01<I, D, M>
where I: WriteOnlyDataCommand, D: DisplayDefinition,

pub fn into_buffered_graphics(self) -> Gc9a01<I, D, BufferedGraphics<D>>

Convert the display into a buffered graphics mode, supporting embedded-graphics.

More information about BufferedGraphics

pub fn init_with_addr_mode( &mut self, delay: &mut impl DelayNs, ) -> Result<(), DisplayError>

Initialise the screen in one of the available addressing modes.

Errors

This method may return an error if there are communication issues with the display.

pub fn draw(&mut self, buffer: &[u8]) -> Result<(), DisplayError>

๐Deprecated: Use draw_buffer instead

Send a raw buffer to the screen.

Errors

This method may return an error if there are communication issues with the display.

pub fn draw_buffer(&mut self, buffer: &[u16]) -> Result<(), DisplayError>

Send a raw buffer to the screen.

Errors

This method may return an error if there are communication issues with the display.

pub fn bounded_draw( &mut self, buffer: &[u16], disp_width: usize, upper_left: (u16, u16), lower_right: (u16, u16), ) -> Result<(), DisplayError>

Send the data to the display for drawing at the current position in the framebuffer and advance the position accordingly. Ref. set_draw_area to modify the affected area by this method.

Notes

This method takes advantage of the bounding box for faster writes. Meaning, it will split into chuncks of write operations.

Errors

This method may return an error if there are communication issues with the display.

pub fn clear_fit_custom_stack<const CLEAR_SIZE_STACK: usize>( &mut self, ) -> Result<(), DisplayError>

Clears the screen by sending a zeroed buffer using a custom stack size for allocation.

This function uses a stack-allocated buffer of size CLEAR_SIZE_STACK to send a portion of the screen-clearing data to the display. The full display area is divided into smaller chunks of size CLEAR_SIZE_STACK, which are then sent iteratively to avoid allocating large amounts of memory on the stack in one go.

This function uses set_draw_area.

Type Parameters

CLEAR_SIZE_STACK: The size of the stack-allocated buffer used to send each chunk of data to the display. This value determines the number of zeroed u16 values sent per iteration.

Errors

This method returns an error if there are communication issues while sending the data to the display.

Panics

This method will stackoverflow if the value of CLEAR_SIZE_STACK exceeds the remaining available stack space.

pub fn clear_fit(&mut self) -> Result<(), DisplayError>

Clears the screen by sending a zeroed buffer using a default stack size.

This method sends zeroed u16 values to the display in chunks to clear the entire screen. It uses a default stack allocation size of 32 u16 values (64 bytes) per iteration, making it suitable for most small embedded contexts where stack space is limited.

If you need to customize the size of the stack-allocated buffer, use [clear_fit_custom_stack] with a specific stack size.

This function uses set_draw_area.

Errors

This method returns an error if there are communication issues while sending the data to the display.

Notes

The default stack allocation size is 64 bytes (32 u16 values).

pub fn set_display_rotation( &mut self, rotation: DisplayRotation, ) -> Result<(), DisplayError>

Set the screen rotation.

Errors

This method may return an error if there are communication issues with the display.

pub fn set_brightness( &mut self, brightness: Brightness, ) -> Result<(), DisplayError>

Change the display brightness.

Errors

This method may return an error if there are communication issues with the display.

pub fn set_screen_state(&mut self, on: Logical) -> Result<(), DisplayError>

Set hardware screen state

Errors

This method may return an error if there are communication issues with the display.

pub fn set_invert_pixels(&mut self, value: bool) -> Result<(), DisplayError>

Set hardware to inverse the GDDRAM framebuffer output

Errors

This method may return an error if there are communication issues with the display.

pub fn set_draw_area( &mut self, start: (u16, u16), end: (u16, u16), ) -> Result<(), DisplayError>

Set hardware framebuffer to configure a limited area of the screen where any pixel should be draw.

(x_start, y_start) - starting point

(x_end, y_end) - ending point

Errors

This method may return an error if there are communication issues with the display.

pub fn set_write_mode(&mut self) -> Result<(), DisplayError>

Set the hardware framebuffer to await incoming colors

Errors

This method may return an error if there are communication issues with the display.

pub const fn get_screen_rotation(&self) -> DisplayRotation

Get screen rotation

pub const fn dimensions(&self) -> (u16, u16)

Get pixel screen dimensions

pub const fn bounds(&self) -> (u16, u16)

Get pixel screen bounds (x-1, y-1)