m5dial_bsp::bsp

Struct Input

pub struct Input<'d> { /* private fields */ }

Digital input.

This driver configures the GPIO pin to be an input. Input drivers read the voltage of their pins and convert it to a logical Level.

Implementations

impl Input<'_>

pub async fn wait_for(&mut self, event: Event)

Wait until the pin experiences a particular [Event].

Example

use esp_hal::gpio::{Event, Input, InputConfig};
let mut input_pin = Input::new(peripherals.GPIO4, InputConfig::default());

input_pin.wait_for(Event::LowLevel).await;

Cancellation

This function is not cancellation-safe.

Calling this function will overwrite previous listen operations for this pin, making it side-effectful.

Dropping the [Future] returned by this function will cancel the wait operation. If the event occurs after the future is dropped, a consequent wait operation will ignore the event.

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub async fn wait_for_high(&mut self)

Wait until the pin is high.

See Self::wait_for for more information.

Example

use esp_hal::gpio::{Event, Input, InputConfig};
let mut input_pin = Input::new(peripherals.GPIO4, InputConfig::default());

input_pin.wait_for_high().await;

pub async fn wait_for_low(&mut self)

Wait until the pin is low.

See Self::wait_for for more information.

Example

use esp_hal::gpio::{Event, Input, InputConfig};
let mut input_pin = Input::new(peripherals.GPIO4, InputConfig::default());

input_pin.wait_for_low().await;

pub async fn wait_for_rising_edge(&mut self)

Wait for the pin to undergo a transition from low to high.

See Self::wait_for for more information.

Example

use esp_hal::gpio::{Event, Input, InputConfig};
let mut input_pin = Input::new(peripherals.GPIO4, InputConfig::default());

input_pin.wait_for_rising_edge().await;

pub async fn wait_for_falling_edge(&mut self)

Wait for the pin to undergo a transition from high to low.

See Self::wait_for for more information.

Example

use esp_hal::gpio::{Event, Input, InputConfig};
let mut input_pin = Input::new(peripherals.GPIO4, InputConfig::default());

input_pin.wait_for_falling_edge().await;

pub async fn wait_for_any_edge(&mut self)

Wait for the pin to undergo any transition, i.e low to high OR high to low.

See Self::wait_for for more information.

Example

use esp_hal::gpio::{Event, Input, InputConfig};
let mut input_pin = Input::new(peripherals.GPIO4, InputConfig::default());

input_pin.wait_for_any_edge().await;

impl<'d> Input<'d>

pub fn new(pin: impl InputPin + 'd, config: InputConfig) -> Input<'d>

Creates a new GPIO input.

The pull parameter configures internal pull-up or pull-down resistors.

Example

The following example configures GPIO5 to read a button press. The example assumes that the button is connected such that the pin is low when the button is pressed.

use esp_hal::{
    delay::Delay,
    gpio::{Input, InputConfig, Level, Pull},
};

fn print_when_pressed(button: &mut Input<'_>, delay: &mut Delay) {
    let mut was_pressed = false;
    loop {
        let is_pressed = button.is_low();
        if is_pressed && !was_pressed {
            println!("Button pressed!");
        }
        was_pressed = is_pressed;
        delay.delay_millis(100);
    }
}

let config = InputConfig::default().with_pull(Pull::Up);
let mut button = Input::new(peripherals.GPIO5, config);
let mut delay = Delay::new();

print_when_pressed(&mut button, &mut delay);

pub fn peripheral_input(&self) -> InputSignal<'d>

Returns a peripheral [input][interconnect::InputSignal] connected to this pin.

The input signal can be passed to peripherals in place of an input pin.

Note that the signal returned by this function is frozen.

Example

use esp_hal::gpio::{Input, InputConfig, Pull};
let config = InputConfig::default().with_pull(Pull::Up);
let pin1_gpio = Input::new(peripherals.GPIO1, config);
let pin1 = pin1_gpio.peripheral_input();

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn is_high(&self) -> bool

Get whether the pin input level is high.

Example

use esp_hal::gpio::{Input, InputConfig};
let pin = Input::new(peripherals.GPIO5, InputConfig::default());
let is_high = pin.is_high();

pub fn is_low(&self) -> bool

Get whether the pin input level is low.

Example

use esp_hal::gpio::{Input, InputConfig};
let pin = Input::new(peripherals.GPIO5, InputConfig::default());
let is_low = pin.is_low();

pub fn level(&self) -> Level

Get the current pin input level.

Example

use esp_hal::gpio::{Input, InputConfig, Level};
let pin = Input::new(peripherals.GPIO5, InputConfig::default());
let level = pin.level();

pub fn apply_config(&mut self, config: &InputConfig)

Change the configuration.

Example

use esp_hal::gpio::{Input, InputConfig, Level, Pull};
let mut pin = Input::new(peripherals.GPIO5, InputConfig::default());
pin.apply_config(&InputConfig::default().with_pull(Pull::Up));

pub fn listen(&mut self, event: Event)

Listen for interrupts.

The interrupts will be handled by the handler set using [Io::set_interrupt_handler]. All GPIO pins share the same interrupt handler.

Note that [Event::LowLevel] and [Event::HighLevel] are fired continuously when the pin is low or high, respectively. You must use a custom interrupt handler to stop listening for these events, otherwise your program will be stuck in a loop as long as the pin is reading the corresponding level.

Examples

Print something when a button is pressed.

use esp_hal::gpio::{Event, Input, InputConfig, Io, Pull};

let mut io = Io::new(peripherals.IO_MUX);
io.set_interrupt_handler(handler);

// Set up the input and store it in the static variable.
// This example uses a push button that is high when not
// pressed and low when pressed.
let config = InputConfig::default().with_pull(Pull::Up);
let mut button = Input::new(peripherals.GPIO5, config);

critical_section::with(|cs| {
    // Here we are listening for a low level to demonstrate
    // that you need to stop listening for level interrupts,
    // but usually you'd probably use `FallingEdge`.
    button.listen(Event::LowLevel);
    BUTTON.borrow_ref_mut(cs).replace(button);
});

// Outside of your `main` function:

use core::cell::RefCell;

use critical_section::Mutex;
use esp_hal::gpio::Input;

// You will need to store the `Input` object in a static variable so
// that the interrupt handler can access it.
static BUTTON: Mutex<RefCell<Option<Input>>> = Mutex::new(RefCell::new(None));

#[handler]
fn handler() {
    critical_section::with(|cs| {
        let mut button = BUTTON.borrow_ref_mut(cs);
        let Some(button) = button.as_mut() else {
            // Some other interrupt has occurred
            // before the button was set up.
            return;
        };

        if button.is_interrupt_set() {
            print!("Button pressed");

            // If you want to stop listening for interrupts, you need to
            // call `unlisten` here. If you comment this line, the
            // interrupt will fire continuously while the button
            // is pressed.
            button.unlisten();
        }
    });
}

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn unlisten(&mut self)

Stop listening for interrupts

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn clear_interrupt(&mut self)

Clear the interrupt status bit for this Pin

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn is_interrupt_set(&self) -> bool

Checks if the interrupt status bit for this Pin is set

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn wakeup_enable( &mut self, enable: bool, event: WakeEvent, ) -> Result<(), WakeConfigError>

Enable as a wake-up source.

This will unlisten for interrupts

Error

Configuring pin to wake up from light sleep on an edge trigger is currently not supported, corresponding variant of [WakeConfigError] will be returned.

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn into_flex(self) -> Flex<'d>

Converts the pin driver into a [Flex] driver.

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.