m5dial_bsp::bsp

Struct M5DialBsp

pub struct M5DialBsp { /* private fields */ }

Holds the board peripherals

Implementations

impl M5DialBsp

pub fn new( bl: Output<'static>, hold: Output<'static>, wake: Input<'static>, ) -> M5DialBsp

pub fn set_backlight(&mut self, state: bool)

Screen backlight control

Arguments:

state: Set backlight ON if true, OFF if false

pub fn shutdown(&mut self)

Shutdown the board.

This method shut-down the board by setting the pin G46 / signal HOLD low. This has no effect when powered by USB. It only works when the board is powered using the green screw terminal (P5 on schematics).

NOTE: The pin signal is set high as start by the init() function.

NOTE: Untested

pub fn is_button_pushed(&mut self) -> bool

Query current button state

pub fn has_button_changed(&mut self) -> Option<bool>

Query if button state has changed since last call.

Returns

Some(current state) if button state has changed.

None if button state has not changed.