m5dial_bsp::bsp

Trait DisplayConfiguration

pub trait DisplayConfiguration<DELAY>
where
    DELAY: DelayNs,
{
    type Error;

    // Required methods
    fn set_rotation(
        &mut self,
        rotation: DisplayRotation,
    ) -> Result<(), Self::Error>;
    fn init(&mut self, delay: &mut DELAY) -> Result<(), Self::Error>;
}

Required Associated Types

type Error

The type representing errors that may occur during display configuration.

Required Methods

fn set_rotation(&mut self, rotation: DisplayRotation) -> Result<(), Self::Error>

Sets the rotation of the display.

Errors

This method may return an error if there are communication issues with the display.

Arguments

rotation - The rotation to set for the display.

Returns

Returns Ok(()) if the rotation is set successfully, otherwise returns an error.

fn init(&mut self, delay: &mut DELAY) -> Result<(), Self::Error>

Initializes and configures the display for the given mode.

Errors

This method may return an error if there are communication issues with the display.

Arguments

delay - A mutable reference to the delay provider used for timing operations.

Returns

Returns Ok(()) if the display is successfully initialized and configured, otherwise returns an error.