Crate m5dial_bsp

M5 Dial Board Support Package (BSP)

This crates is a Board support package for the M5 Dial.

API Documenation

Feature list/roadmap:

Screen driver (GC9A01)

Rotary encoder

Button

Power management (Shutdown)

Touch-screen (FT3267)

Buzzer

Real-time clock (BM8563)

Port A (I2C)

Port B

How to use

First generate a base project using esp-generate or esp-idf-template as described in they respective documentations. Then, add this crate to your Cargo.toml:

[dependencies]
....
rotary-encoder-hal = "0.6.0"
....

In your main function then initialize this hall and use it:

let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
let peripherals = esp_hal::init(config);

let mut board = m5dial::init(peripherals);
....

For more information, please refer to the API Documenation.

Examples

See screen_counter.rs

License

MIT license (LICENSE or http://opensource.org/licenses/MIT)

Contriutions

… are welcome !

Credit

This crates rely on many crates of the embedded Rust working group, ESP support crates from Expressif, embedded-graphics and device drivers such as :

rotary-encoder-hal

gc9c01 display driver

Re-exports

pub use bsp::init;

pub use bsp::M5DialBsp;

Modules

bsp

M5Dial Board Support Package